+++ fs/src/lib.rs
@@ -39,6 +39,8 @@ pub mod stat;
 pub use btrfs_disk::{items::DeviceItem, superblock::Superblock};
 pub use cache::{CacheStats, LruTreeBlockCache};
 pub use dir::{Entry, FileKind};
-pub use filesystem::{Filesystem, Inode, StatFs, SubvolId, SubvolInfo};
+pub use filesystem::{
+    Filesystem, Inode, SearchFilter, SearchItem, StatFs, SubvolId, SubvolInfo,
+};
 pub use stat::Stat;
 pub use uuid::Uuid;